feat: Add statistics endpoints
Adds endpoints for statistics, including listing, adding and deleting
statistics. This commit also includes k6 tests for the new endpoints
and updates the necessary configurations and models.

In addition this fixes a lot of warnings in the codebase, ensuring
better code quality and maintainability.

Adds multiple new error types to the `apperror` module to handle
statistics-related errors more effectively.

Breaking changes: None

Resolves: None

+++ src/model/config.rs
@@ -44,7 +44,7 @@ pub struct Config {
 #[serde(rename_all = "camelCase")]
 pub struct Database {
     /**
-     * Type of the database (e.g., PostgreSQL).
+     * Type of the database (e.g., `PostgreSQL`).
      */
     pub db_type: DatabaseType,
 }
@@ -53,10 +53,10 @@ pub struct Database {
 #[serde(rename_all = "camelCase")]
 pub enum DatabaseType {
     /**
-     * PostgreSQL database type.
+     * `PostgreSQL` database type.
      */
     #[serde(rename_all = "camelCase")]
-    Postgresql { connection_string: String, max_connections: usize, min_connections: usize, acquire_timeout: usize, acquire_slow_threshold: usize, idle_timeout: usize, max_lifetime: usize },
+    Postgresql { connection_string: String, max_connections: u32, min_connections: u32, acquire_timeout: u64, acquire_slow_threshold: u64, idle_timeout: u64, max_lifetime: u64 },
 }
 
 #[derive(Clone, Serialize, Deserialize)]
